requirements.txt
aiofiles==23.1.0
asgiref==3.6.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.6
dict-toolbox==3.0.3
Django==4.1.6
enum34==1.1.10
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
msgpack==1.0.5
nest-asyncio==1.5.6
packaging==21.3
pluggy==1.0.0
pop==23.1.0
pop-config==12.0.1
pop-loop==1.0.6
py==1.11.0
pydevd-pycharm==212.5457.59
pyparsing==3.0.9
pyTelegramBotAPI==0.3.0
pytest==7.1.2
python-decouple==3.8
PyYAML==6.0
rend==6.5.2
requests==2.27.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
termcolor==2.2.0
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.9
